# docs/generated/openapi.json
{
  "components": {
    "schemas": {
      "AnswerIn": {
        "properties": {
          "question": {
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "top_k": {
            "default": 5,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Top K",
            "type": "integer"
          }
        },
        "required": [
          "question"
        ],
        "title": "AnswerIn",
        "type": "object"
      },
      "AnswerOut": {
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "title": "Citations",
            "type": "array"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "answered",
              "insufficient_evidence"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "answer",
          "citations",
          "request_id"
        ],
        "title": "AnswerOut",
        "type": "object"
      },
      "Citation": {
        "properties": {
          "chunk_id": {
            "title": "Chunk Id",
            "type": "string"
          },
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "source_text": {
            "title": "Source Text",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "chunk_id",
          "title",
          "source_text",
          "rank",
          "score"
        ],
        "title": "Citation",
        "type": "object"
      },
      "DocumentIn": {
        "properties": {
          "allowed_groups": {
            "items": {
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 100,
            "title": "Allowed Groups",
            "type": "array"
          },
          "document_id": {
            "minLength": 1,
            "title": "Document Id",
            "type": "string"
          },
          "owner_subject": {
            "minLength": 1,
            "title": "Owner Subject",
            "type": "string"
          },
          "text": {
            "minLength": 1,
            "title": "Text",
            "type": "string"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "version": {
            "minLength": 1,
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "version",
          "title",
          "text",
          "owner_subject"
        ],
        "title": "DocumentIn",
        "type": "object"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorOut": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
      "IngestOut": {
        "properties": {
          "chunk_count": {
            "title": "Chunk Count",
            "type": "integer"
          },
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "version",
          "chunk_count",
          "request_id"
        ],
        "title": "IngestOut",
        "type": "object"
      },
      "SearchHit": {
        "properties": {
          "chunk_id": {
            "title": "Chunk Id",
            "type": "string"
          },
          "dense_score": {
            "title": "Dense Score",
            "type": "number"
          },
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "fused_score": {
            "title": "Fused Score",
            "type": "number"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "sparse_score": {
            "title": "Sparse Score",
            "type": "number"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "chunk_id",
          "title",
          "text",
          "rank",
          "sparse_score",
          "dense_score",
          "fused_score"
        ],
        "title": "SearchHit",
        "type": "object"
      },
      "SearchIn": {
        "properties": {
          "query": {
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "top_k": {
            "default": 5,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Top K",
            "type": "integer"
          }
        },
        "required": [
          "query"
        ],
        "title": "SearchIn",
        "type": "object"
      },
      "SearchOut": {
        "properties": {
          "hits": {
            "items": {
              "$ref": "#/components/schemas/SearchHit"
            },
            "title": "Hits",
            "type": "array"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "hits",
          "request_id"
        ],
        "title": "SearchOut",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "RAG Engineering API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Health Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "死活状態を取得する",
        "tags": [
          "system"
        ]
      }
    },
    "/v1/answers": {
      "post": {
        "description": "認可済み根拠だけで回答し、根拠不足時は明示的に回答を拒否します。",
        "operationId": "answer_v1_answers_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-principal-groups",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Principal-Groups"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "認証情報がない、または形式が不正です。"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "操作または根拠への権限がありません。"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "入力が OpenAPI schema に適合しません。"
          }
        },
        "summary": "引用付き回答を生成する"
      }
    },
    "/v1/documents": {
      "post": {
        "description": "版と ACL を保持して文書を正規化、チャンク化、索引化します。",
        "operationId": "ingest_document_v1_documents_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-principal-groups",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Principal-Groups"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "認証情報がない、または形式が不正です。"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "操作または根拠への権限がありません。"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "入力が OpenAPI schema に適合しません。"
          }
        },
        "summary": "文書を取り込む"
      }
    },
    "/v1/search": {
      "post": {
        "description": "ACL hard filter 後に疎密ハイブリッド検索と順位融合を行います。",
        "operationId": "search_v1_search_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-principal-groups",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Principal-Groups"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "認証情報がない、または形式が不正です。"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "操作または根拠への権限がありません。"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "入力が OpenAPI schema に適合しません。"
          }
        },
        "summary": "認可済み根拠を検索する"
      }
    }
  }
}